split the controller
