+++ routes/userRoutes.js
@@ -1,6 +1,10 @@
 const express = require("express");
 const userRouter = express();
-const userController = require("../controllers/userController");
+// const userController = require("../controllers/userController");
+const authController = require("../controllers/authController");
+const profileController = require("../controllers/profileController");
+const viewController = require("../controllers/viewController");
+
 const Auth = require("../middleware/authMiddleware");
 const upload = require("../middleware/multer");
 const CourseController = require("../controllers/courseController");
@@ -10,48 +14,42 @@ const {
   findChats,
 } = require("../controllers/chatController");
 const {
-createMessage,
+  createMessage,
   getMessage,
 } = require("../controllers/messageController");
-userRouter.post("/signup", userController.adduser);
-userRouter.post("/sendotp", userController.sendOTP);
-userRouter.post("/verifyotp", userController.userverifyOTP);
-userRouter.post("/login", userController.verifyLogin);
-userRouter.post("/forgotPass", userController.forgotPass);
-userRouter.post("/newPass", userController.passverifyOTP);
-userRouter.put("/updateaPass", userController.updatePass);
-userRouter.post("/userRegisterWithGoole", userController.googleRegister);
-userRouter.get("/getUser", userController.getUser);
-userRouter.post(
-  "/updateProfile",
-  upload.single("image"),
-  userController.UpdateProfile
-);
+
+userRouter.post("/signup", authController.adduser);
+userRouter.post("/sendotp", authController.sendOTP);
+userRouter.post("/verifyotp", authController.userverifyOTP);
+userRouter.post("/login", authController.verifyLogin);
+userRouter.post("/forgotPass", authController.forgotPass);
+userRouter.post("/newPass", authController.passverifyOTP);
+userRouter.put("/updateaPass", authController.updatePass);
+userRouter.post("/userRegisterWithGoole", authController.googleRegister);
+userRouter.get("/getUser", profileController.getUser);
+userRouter.post("/updateProfile",upload.single("image"),profileController.UpdateProfile);
 
 //  Get
-userRouter.get("/manageProfile", userController.manageProfile);
+userRouter.get("/manageProfile", profileController.manageProfile);
 userRouter.get("/getCourse", CourseController.getCourse);
-userRouter.get("/LoadCategory", userController.getCategory);
-userRouter.get("/fetchReview", userController.fetchReview);
+userRouter.get("/LoadCategory", CourseController.getCategory);
+userRouter.get("/fetchReview",viewController.fetchReview);
 userRouter.get("/findUserChats/:userId", findUserChats);
 userRouter.get("/findChat/:firstId/:secondId", findChats);
 userRouter.get("/getMeassage/:chatId", getMessage);
-userRouter.get("/getCommnets/:id",userController.getCommnets);
+userRouter.get("/getCommnets/:id", viewController.getCommnets);
 
 // Post
-userRouter.post("/updateuser", Auth, userController.profileUpdate);
-userRouter.post("/enrollments", userController.enrollments);
-userRouter.post("/checkout", userController.checkout);
-userRouter.post("/addReview", userController.addReview);
+userRouter.post("/updateuser", Auth, profileController.profileUpdate);
+userRouter.post("/enrollments", CourseController.enrollments);
+userRouter.post("/checkout", CourseController.checkout);
+userRouter.post("/addReview", viewController.addReview);
 userRouter.post("/createChat", createChat);
 userRouter.post("/createMessage", createMessage);
-userRouter.post("/postCommnets", userController.postCommnets);
-userRouter.post("/createBlog",upload.single("image"), userController.createBlog);
-
+userRouter.post("/postCommnets", viewController.postCommnets);
+userRouter.post("/createBlog",upload.single("image"),viewController.createBlog);
 
 // Put
-userRouter.put("/purchaseCourse/:id", Auth, userController.purchaseCourse);
-
-
+userRouter.put("/purchaseCourse/:id", Auth, CourseController.purchaseCourse);
 
 module.exports = userRouter;
